+++ src/components/relatedTour/index.jsx
@@ -1,10 +1,11 @@
 import React from "react";
 import radium, { Style } from "radium";
-import { color, media } from "../../../settings.json";
+import { media } from "../../../settings.json";
+import color from "../../styles/colors";
 import Heading from "../heading";
 import Price from "../price";
 import Bullet from "../bullet";
-import { gutter, span, percentage } from "../../utils/grid";
+import { span } from "../../utils/grid";
 
 const styles = {
   base: {},
@@ -13,9 +14,14 @@ const styles = {
     image: {
       base: {
         display: "block",
-        marginBottom: gutter("static"),
+        marginBottom: "32px",
         width: "100%",
       },
+    },
+
+    heading: {
+      fontSize: "28px",
+      lineHeight: 1,
     },
 
     content: {
@@ -26,17 +32,20 @@ const styles = {
 
     details: {
       base: {
-        color: color.text,
+        color: color.textPrimary,
         fontSize: "1.1rem",
         lineHeight: 1,
         textTransform: "uppercase",
 
         [`@media (max-width: ${media.max["480"]})`]: {
           display: "inline-block",
+          color: color.accentGray,
+          marginTop: "14px",
+          fontWeight: 600,
         },
 
         [`@media (min-width: ${media.min["480"]})`]: {
-          marginTop: ".7rem",
+          marginTop: "1.4rem",
         },
       },
 
@@ -116,35 +125,38 @@ function RelatedTour({
           },
 
           ".Price-amount": {
-            color: color.red,
-            fontSize: "1.1rem",
+            color: color.textPrimary,
+            fontSize: "2.4rem !important",
           },
 
           mediaQueries: {
             [`(max-width: ${media.max["480"]})`]: {
               ".Price": {
-                display: "inline-block",
+                display: "block",
+                marginTop: "20px",
+
               },
 
               ".Price-amount": {
-                fontWeight: 600,
+                fontWeight: "300 !important",
+                color: color.accentRed,
               },
             },
 
             [`(min-width: ${media.min["480"]})`]: {
               ".Heading": {
                 maxWidth: "50rem",
-                width: percentage("50rem", span(6, "static")),
+                width: "80%",
               },
 
               ".Price": {
-                position: "absolute",
+                position: "absolute !important",
                 right: 0,
-                top: ".8rem",
+                top: 0,
               },
 
               ".Price-amount": {
-                color: `${color.darkGray}`,
+                color: color.textPrimary,
                 fontSize: "2.4rem",
               },
             },
@@ -175,18 +187,20 @@ function RelatedTour({
           weight="thick"
           truncate
           tracking="tight"
+          override={styles.elements.heading}
         >
           <a href={slug} onClick={clickHandler}>
             {title}
           </a>
         </Heading>
 
+        {Details}
+
         <Price
           amount={price.amount}
           currency={price.currency}
         />
 
-        {Details}
       </div>
     </div>
   );