+++ src/styles/typography.js
@@ -1,6 +1,8 @@
 export const fontWeightLight = 300;
-export const fontWeightBook = 400;
+export const fontWeightRegular = 500;
 export const fontWeightMedium = 600;
+// Book is deprecated and will be removed in the next major release
+export const fontWeightBook = fontWeightRegular;
 
 export const fontSizeSuper = 88;
 export const fontSizeHeading1 = 64;
@@ -10,20 +12,24 @@ export const fontSizeHeading4 = 32;
 export const fontSizeHeading5 = 24;
 export const fontSizeHeading6 = 20;
 export const fontSizeHeading7 = 16;
+export const fontSizeHeading8 = 11;
 export const fontSizeBodyArticle = 20;
+export const fontSizeBodyArticleSmall = 16;
 export const fontSizeBodySmall = 14;
 export const fontSizeAccent = 16;
 export const fontSizeUppercase = 11;
 
 export const lineHeightSuper = (96 / fontSizeSuper);
-export const lineHeightHeading1 = (76 / fontSizeHeading1);
+export const lineHeightHeading1 = (72 / fontSizeHeading1);
 export const lineHeightHeading2 = (56 / fontSizeHeading2);
 export const lineHeightHeading3 = (48 / fontSizeHeading3);
 export const lineHeightHeading4 = (40 / fontSizeHeading4);
-export const lineHeightHeading5 = (28 / fontSizeHeading5);
-export const lineHeightHeading6 = (24 / fontSizeHeading6);
-export const lineHeightHeading7 = (20 / fontSizeHeading7);
+export const lineHeightHeading5 = (32 / fontSizeHeading5);
+export const lineHeightHeading6 = (28 / fontSizeHeading6);
+export const lineHeightHeading7 = (24 / fontSizeHeading7);
+export const lineHeightHeading8 = (16 / fontSizeHeading8);
 export const lineHeightBodyArticle = (36 / fontSizeBodyArticle);
+export const lineHeightBodyArticleSmall = (32 / fontSizeBodyArticleSmall);
 export const lineHeightBodySmall = (20 / fontSizeBodySmall);
 export const lineHeightAccent = (24 / fontSizeAccent);
 export const lineHeightUppercase = (11 / fontSizeUppercase);